refactor: 졸 기물, 전진에 대한 분기 처리를 Enum으로 하도록 수정
