+++ src/main/java/team/janggi/domain/piece/Piece.java
@@ -6,13 +6,12 @@ import team.janggi.domain.Team;
 import team.janggi.domain.board.BoardStateReader;
 import team.janggi.domain.piece.strategy.CannonMoveStrategy;
 import team.janggi.domain.piece.strategy.ChariotPalaceMoveStrategy;
-import team.janggi.domain.piece.strategy.royal.ChoRoyalPieceMoveStrategy;
-import team.janggi.domain.piece.strategy.ChoSoldierMoveStrategy;
 import team.janggi.domain.piece.strategy.ElephantMoveStrategy;
-import team.janggi.domain.piece.strategy.royal.HanRoyalPieceMoveStrategy;
-import team.janggi.domain.piece.strategy.HanSoldierMoveStrategy;
 import team.janggi.domain.piece.strategy.HorseMoveStrategy;
 import team.janggi.domain.piece.strategy.MoveStrategy;
+import team.janggi.domain.piece.strategy.SoldierMoveStrategy;
+import team.janggi.domain.piece.strategy.royal.ChoRoyalPieceMoveStrategy;
+import team.janggi.domain.piece.strategy.royal.HanRoyalPieceMoveStrategy;
 
 public class Piece {
     private final Team team;
@@ -59,10 +58,10 @@ public class Piece {
 
     private static MoveStrategy getSoldierMoveStrategyByTeam(Team team) {
         if (team == Team.CHO) {
-            return ChoSoldierMoveStrategy.instance;
+            return SoldierMoveStrategy.towardTopInstance;
         }
         if (team == Team.HAN) {
-            return HanSoldierMoveStrategy.instance;
+            return SoldierMoveStrategy.towardBottomInstance;
         }
         throw new IllegalArgumentException("졸은 초 또는 한 팀에 속해야 합니다.");
     }
