+++ src/main/java/team/janggi/domain/piece/strategy/ForwardDirection.java
@@ -1,0 +1,6 @@
+package team.janggi.domain.piece.strategy;
+
+public enum ForwardDirection {
+    TOWARD_TOP,
+    TOWARD_BOTTOM
+}
