+++ src/main/java/team/janggi/domain/piece/strategy/SoldierPalaceMoveStragtegy.java
@@ -1,0 +1,4 @@
+package team.janggi.domain.piece.strategy;
+
+public class SoldierPalaceMoveStragtegy {
+}
